+++ app/src/main/kotlin/com/simplemobiletools/calendar/helpers/CalDAVHandler.kt
@@ -10,6 +10,7 @@ import android.util.SparseIntArray
 import com.simplemobiletools.calendar.activities.SimpleActivity
 import com.simplemobiletools.calendar.extensions.config
 import com.simplemobiletools.calendar.extensions.dbHelper
+import com.simplemobiletools.calendar.extensions.refreshCalDAVCalendars
 import com.simplemobiletools.calendar.extensions.scheduleCalDAVSync
 import com.simplemobiletools.calendar.models.CalDAVCalendar
 import com.simplemobiletools.calendar.models.Event
@@ -188,7 +189,9 @@ class CalDAVHandler(val context: Context) {
 
         var sortedColors = ArrayList<Int>(colors.size())
         (0 until colors.size()).mapTo(sortedColors) { colors[it] }
-        sortedColors = sortedColors.distinct() as ArrayList<Int>
+        if (sortedColors.isNotEmpty()) {
+            sortedColors = sortedColors.distinct() as ArrayList<Int>
+        }
 
         return sortedColors
     }
@@ -322,6 +325,7 @@ class CalDAVHandler(val context: Context) {
 
         setupCalDAVEventReminders(event)
         setupCalDAVEventImportId(event)
+        refreshCalDAVCalendar(event)
     }
 
     fun updateCalDAVEvent(event: Event) {
@@ -335,6 +339,7 @@ class CalDAVHandler(val context: Context) {
 
         setupCalDAVEventReminders(event)
         setupCalDAVEventImportId(event)
+        refreshCalDAVCalendar(event)
     }
 
     private fun setupCalDAVEventReminders(event: Event) {
@@ -412,12 +417,14 @@ class CalDAVHandler(val context: Context) {
             context.contentResolver.delete(contentUri, null, null)
         } catch (ignored: Exception) {
         }
+        refreshCalDAVCalendar(event)
     }
 
     fun insertEventRepeatException(event: Event, occurrenceTS: Int): Long {
         val uri = CalendarContract.Events.CONTENT_URI
         val values = fillEventRepeatExceptionValues(event, occurrenceTS)
         val newUri = context.contentResolver.insert(uri, values)
+        refreshCalDAVCalendar(event)
         return java.lang.Long.parseLong(newUri.lastPathSegment)
     }
 
@@ -458,4 +465,6 @@ class CalDAVHandler(val context: Context) {
     }
 
     private fun getCalDAVEventImportId(calendarId: Int, eventId: Long) = "$CALDAV-$calendarId-$eventId"
+
+    private fun refreshCalDAVCalendar(event: Event) = context.refreshCalDAVCalendars(null, event.getCalDAVCalendarId().toString())
 }